chore: refactor adj communities to separate file
