+++ packages/hierarchy-browser/src/CommunityCard/AdjacentCommunities/index.tsx
@@ -1,0 +1,1 @@
+export { AdjacentCommunities } from './AdjacentCommunities'
